Add clarification about index and length of string
Adding a note that explains that the index of a string is actually a byte of that string, and the length of a string is the length in bytes; not characters which would only technically work for ASCII characters in the string.  A link is provided to the applicable blog post.