Move <img> for play.golang.org to separate line in the template
This will make future diffs easier to read because code changes modify the
hash/link, but the image data stays the same.

Note that there are whitespace diffs in the generated files where there
is no image. This is because there was a stray tab in the template before,
and now it's replaced by spaces.

Fixes #235