don't embed the mutex
Embedding a mutex is a really bad idea. It exposes the Lock and Unlock methods on the enclosing type, which then allows code from outside your type to mess with your mutex.

There was no reason to embed the mutex in this example. The comment say it is idiomatic to do so, but that is untrue.  Using a mutex called "mu" is idiomatic.